added non number id error and checks

+++ project/wallet-full-stack-app/backend/server.js
@@ -48,7 +48,9 @@ app.get(`/api/transactions/:userId`, async (req, res) => {
 app.delete("/api/transactions/:id", async (req, res) => {
   try {
     const { id } = req.params;
-
+    if(isNaN(id)){
+      return res.status(400).json({ message: "Invalid transaction ID" });
+    }
     const result = await sql`
     DELETE FROM transactions WHERE id = ${id} RETURNING *`;
